+++ js/utils/get-random-positive-float.js
@@ -20,3 +20,5 @@ function getRandomPositiveFloat (a, b, digits = 1) {
   // указать требуемое количество знаков после точки
   return result.toFixed(digits);
 }
+
+export {getRandomPositiveFloat};
